Controllers edited

+++ PhotoAlbomAPI/API/Controllers/AccountManagerController.cs
@@ -2,13 +2,14 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using API.Models;
 using API.Models.ViewModels;
 using BLL.Interfaces;
 using BLL.Models;
-using DAL.Entities;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 namespace API.Controllers
 {
@@ -18,13 +19,15 @@ namespace API.Controllers
     {
 
         private readonly IAccountManagerService<UserBLL> _accountManagerService;
+        private readonly IOrganisationService<OrganisationBLL> _organisationService;
 
-        public AccountManagerController(IAccountManagerService<UserBLL> accountManagerService)
+        public AccountManagerController(IAccountManagerService<UserBLL> accountManagerService, IOrganisationService<OrganisationBLL> organisationService)
         {
             _accountManagerService = accountManagerService;
+            _organisationService = organisationService;
         }
 
-        [HttpGet]
+        [HttpGet("users")]
         public async Task<IEnumerable<UserViewModel>> GetAllUsers()
         {
             var users = await _accountManagerService.GetAllUsers();
@@ -35,8 +38,18 @@ namespace API.Controllers
             }
             return userVM;
         }
-
-        [HttpPost("create")]
+        [HttpGet("organisation")]
+        public async Task<IEnumerable<OrganisationModel>> GetAllOrganisation()
+        {
+            var orgs = await _organisationService.GetAllAsync();
+            var orgsModel = new List<OrganisationModel>();
+            foreach (var item in orgs)
+            {
+                orgsModel.Add(new OrganisationModel { Id = item.Id, Name = item.Name, Location = item.Location });
+            }
+            return orgsModel;
+        }
+        [HttpPost("create")]    
         public async Task<IActionResult> CreacteAccount([FromBody]AccountCreateViewModel model)
         {
             if (!ModelState.IsValid)
@@ -60,7 +73,7 @@ namespace API.Controllers
             }
         }
 
-        [HttpPut("edit")]
+        [HttpPut("users/edit")]
         public async Task<IActionResult> EditUserAccount([FromBody] EditUserNameViewModel model)
         {
             if (ModelState.IsValid)
@@ -82,15 +95,75 @@ namespace API.Controllers
             return BadRequest(ModelState);
         }
 
-        [HttpDelete("delete")]
-        public async Task<ActionResult> DeleteAccount([FromBody] DeleteUserViewModel model)
+        [HttpDelete("users/delete")]
+        public async Task<ActionResult> DeleteUserAccount([FromBody] DeleteUserViewModel model)
         {
-            var result = await _accountManagerService.DeleteAccount(model.UserName);
-            if (result.Succeeded)
+            if (ModelState.IsValid)
             {
-                return new OkObjectResult("Account deleted");
+                var result = await _accountManagerService.DeleteAccount(model.UserName);
+                if (result.Succeeded)
+                {
+                    return new OkObjectResult("Account deleted");
+                }
+                return NotFound("Invalid user name");
             }
-            return NotFound("Invalid user name");
+            return BadRequest(ModelState);
+        }
+
+        [HttpPut("organisation/edit")]
+        public async Task<IActionResult> EditOrganisationAccount([FromBody] EditOrganisationViewModel model)
+        {
+
+            //if (!ModelState.IsValid)
+            //{
+            //    return BadRequest("Not a valid model");
+            //}
+            //if (_organisationService.GetAsync(model.Id) == null)
+            //{
+            //    return BadRequest("Model doesn`t exicte");
+            //}
+            //await _organisationService.UpdateAsync(new OrganisationBLL { Id= model.Id, Name = model.OrgName, Location = model.Loacation});
+
+            //return new OkObjectResult("Account updated");
+
+            if (ModelState.IsValid)
+            {
+              
+
+                try
+                {
+                    await _organisationService.UpdateAsync(new OrganisationBLL { Id = model.Id, Name = model.OrgName, Location = model.Loacation });
+                }
+                catch (DbUpdateConcurrencyException)
+                {
+                    if (await _organisationService.GetAsync(model.Id) == null)
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
+                }
+                return new OkObjectResult(model);
+            }
+            return BadRequest("Not a valid model");
+        }
+
+        [HttpDelete("organisation/delete")]
+        public async Task<ActionResult> DeleteOrganisationAccount([FromBody]DeleteOrganisationViewModel model)
+        {
+            if (ModelState.IsValid)
+            {
+                await _organisationService.DeleteAsync(model.OrgId);
+                var result = await _accountManagerService.DeleteAccount(model.UserName);
+                if (result.Succeeded)
+                {
+                    return new OkObjectResult("Account deleted");
+                }
+                return NotFound("Invalid user name");
+            }
+            return BadRequest(ModelState);
         }
     }
 }
